# dead store elimination