Improve inner loop in `work` functions. Also resolves a warning.