@wirew0rm 2023-03-02
Benchmark results:

### main - bm_case1

```
┌──────────────────────────────benchmark:───────────────────────────────┬──────┬─#N──┬─CTX-SW─┬──CPU cache misses───┬──CPU branch misses──┬─<CPU-I>─┬──min───┬──mean──┬─stddev─┬─median─┬──max───┬─total time─┬─ops/s──┐
│ merged src->sink work                                                 │ PASS │ 10  │     1  │   3k /   7k = 38.3% │ 650  / 323k =  0.2% │  348.7m │ 318 us │ 644 us │ 218 us │ 786 us │ 855 us │       6 ms │   1.6G │
│ merged src->copy->sink                                                │ PASS │ 10  │     1  │   2k /   5k = 45.5% │ 409  / 320k =  0.1% │  784.8m │ 326 us │ 338 us │   9 us │ 342 us │ 357 us │       3 ms │   3.0G │
│ merged src->copy->sink work                                           │ PASS │ 10  │     2  │   5k /  11k = 46.9% │ 893  / 325k =  0.3% │  880.8m │ 337 us │ 341 us │   6 us │ 340 us │ 355 us │       3 ms │   2.9G │
│ merged src->copy^10->sink                                             │ PASS │ 10  │     1  │   1k /   4k = 31.4% │ 314  / 319k =  0.1% │  784.4m │ 319 us │ 336 us │  17 us │ 335 us │ 364 us │       3 ms │   3.0G │
│ merged src(N=1024)->b1(N≤128)->b2(N=1024)->b3(N=32...128)->sink       │ PASS │ 10  │     1  │   2k /   6k = 40.4% │ 482  / 321k =  0.2% │  285.2m │ 297 us │ 320 us │  10 us │ 325 us │ 328 us │       3 ms │   3.1G │
│ merged src->mult(2.0)->divide(2.0)->add(-1)->sink - float             │ PASS │ 10  │     2  │   2k /   6k = 30.1% │ 464  / 948k =  0.0% │    1.9  │ 534 us │ 540 us │   4 us │ 541 us │ 547 us │       5 ms │   1.9G │
│ merged src->mult(2.0)->divide(2.0)->add(-1)->sink - int               │ PASS │ 10  │     4  │   6k /  16k = 38.4% │   1k /   6M =  0.0% │    8.8  │   1 ms │   1 ms │   6 us │   1 ms │   1 ms │      11 ms │ 919.9M │
│ merged src->(mult(2.0)->div(2.0)->add(-1))^10->sink - float           │ PASS │ 10  │     8  │   5k /  20k = 28.0% │   1k /   7M =  0.0% │   16.3  │   2 ms │   2 ms │  15 us │   2 ms │   3 ms │      25 ms │ 401.7M │
│ merged src->(mult(2.0)->div(2.0)->add(-1))^10->sink - int             │ PASS │ 10  │    26  │   8k /  50k = 16.8% │   3k /  57M =  0.0% │   84.4  │   7 ms │   7 ms │  22 us │   7 ms │   7 ms │      70 ms │ 142.9M │
│ runtime   src->sink overhead                                          │ PASS │ 10  │     3  │  31k /   2M =  1.6% │   2k /  11M =  0.0% │    7.9  │ 839 us │ 860 us │  23 us │ 855 us │ 920 us │       9 ms │   1.2G │
│ runtime   src->copy->sink                                             │ PASS │ 10  │     3  │ 138k /   4M =  3.7% │   2k /  11M =  0.0% │    8.3  │ 903 us │ 924 us │  35 us │ 912 us │   1 ms │       9 ms │   1.1G │
│ runtime   src->copy^10->sink                                          │ PASS │ 10  │     8  │   2M /  26M =  8.0% │   8k /  14M =  0.1% │   11.8  │   2 ms │   2 ms │ 234 us │   2 ms │   2 ms │      17 ms │ 577.2M │
│ runtime   src(N=1024)->b1(N≤128)->b2(N=1024)->b3(N=32...128)->sink    │ PASS │ 10  │     6  │ 464k /   6M =  7.9% │  76k /  23M =  0.3% │   15.1  │   2 ms │   2 ms │  39 us │   2 ms │   2 ms │      18 ms │ 546.6M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - float             │ PASS │ 10  │     5  │ 240k /  10M =  2.5% │   3k /  13M =  0.0% │   12.7  │   1 ms │   1 ms │  71 us │   1 ms │   2 ms │      15 ms │ 671.5M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - int               │ PASS │ 10  │     7  │ 205k /  10M =  2.1% │   4k /  13M =  0.0% │   17.0  │   2 ms │   2 ms │  73 us │   2 ms │   2 ms │      18 ms │ 547.8M │
│ runtime   src->(mult(2.0)->div(2.0)->add(-1))^10->sink - float        │ PASS │ 10  │    25  │   2M /  77M =  2.2% │  16k /  30M =  0.1% │   55.9  │   7 ms │   7 ms │ 407 us │   7 ms │   8 ms │      73 ms │ 137.6M │
│ runtime   src->(mult(2.0)->div(2.0)->add(-1))^10->sink - int          │ PASS │ 10  │    42  │   2M /  77M =  2.3% │  18k /  31M =  0.1% │   98.4  │  11 ms │  11 ms │ 393 us │  11 ms │  12 ms │     107 ms │  93.8M │
│ runtime   src->mult(2.0)->mult(0.5)->add(-1)->sink (SIMD)             │ PASS │ 10  │     4  │ 242k /  10M =  2.4% │   3k /  18M =  0.0% │   12.4  │   1 ms │   1 ms │  46 us │   1 ms │   1 ms │      12 ms │ 844.3M │
│ runtime   src->(mult(2.0)->mult(0.5)->add(-1))^10->sink (SIMD)        │ PASS │ 10  │    18  │   2M /  78M =  2.8% │  20k /  87M =  0.0% │   53.4  │   4 ms │   4 ms │ 415 us │   4 ms │   5 ms │      42 ms │ 235.8M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - float single      │ PASS │ 10  │     7  │ 233k /  10M =  2.4% │   4k /  13M =  0.0% │   12.7  │   2 ms │   2 ms │ 176 us │   2 ms │   2 ms │      20 ms │ 499.6M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - int single        │ PASS │ 10  │     7  │ 208k /  10M =  2.1% │   4k /  13M =  0.0% │   17.0  │   2 ms │   2 ms │  69 us │   2 ms │   2 ms │      18 ms │ 545.7M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - float bulk        │ PASS │ 10  │     6  │ 250k /  10M =  2.6% │   4k /  18M =  0.0% │   12.5  │   1 ms │   1 ms │  71 us │   1 ms │   2 ms │      13 ms │ 759.8M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - int bulk          │ PASS │ 10  │    10  │ 211k /  10M =  2.2% │   5k /  26M =  0.0% │   19.5  │   3 ms │   3 ms │  79 us │   3 ms │   3 ms │      27 ms │ 376.3M │
└───────────────────────────────────────────────────────────────────────┴──────┴─────┴────────┴─────────────────────┴─────────────────────┴─────────┴────────┴────────┴────────┴────────┴────────┴────────────┴────────┘
```
### main - bm_case1_nosimd

```
┌──────────────────────────────benchmark:───────────────────────────────┬──────┬─#N──┬─CTX-SW─┬──CPU cache misses───┬──CPU branch misses──┬─<CPU-I>─┬──min───┬──mean──┬─stddev─┬─median─┬──max───┬─total time─┬─ops/s──┐
│ merged src->sink work                                                 │ PASS │ 10  │     2  │   2k /   7k = 35.2% │ 694  /  10M =  0.0% │    8.0  │ 497 us │ 505 us │   8 us │ 506 us │ 516 us │       5 ms │   2.0G │
│ merged src->copy->sink                                                │ PASS │ 10  │     2  │   6k /  11k = 53.2% │   1k /  10M =  0.0% │    7.0  │ 497 us │ 505 us │   9 us │ 505 us │ 524 us │       5 ms │   2.0G │
│ merged src->copy->sink work                                           │ PASS │ 10  │     2  │   4k /   9k = 43.6% │ 576  /  10M =  0.0% │    8.0  │ 497 us │ 514 us │   7 us │ 514 us │ 523 us │       5 ms │   1.9G │
│ merged src->copy^10->sink                                             │ PASS │ 10  │     2  │   5k /  11k = 44.8% │ 805  /  10M =  0.0% │    7.0  │ 497 us │ 511 us │  11 us │ 514 us │ 534 us │       5 ms │   2.0G │
│ merged src(N=1024)->b1(N≤128)->b2(N=1024)->b3(N=32...128)->sink       │ PASS │ 10  │     2  │   3k /   7k = 37.6% │ 579  /  10M =  0.0% │    7.0  │ 497 us │ 504 us │   9 us │ 499 us │ 527 us │       5 ms │   2.0G │
│ merged src->mult(2.0)->divide(2.0)->add(-1)->sink - float             │ PASS │ 10  │     6  │   2k /  12k = 16.2% │ 884  /  10M =  0.0% │   12.0  │   1 ms │   1 ms │  14 us │   1 ms │   1 ms │      11 ms │ 942.0M │
│ merged src->mult(2.0)->divide(2.0)->add(-1)->sink - int               │ PASS │ 10  │     6  │   3k /  12k = 22.9% │ 828  /  10M =  0.0% │   13.0  │   1 ms │   2 ms │  20 us │   2 ms │   2 ms │      15 ms │ 660.4M │
│ merged src->(mult(2.0)->div(2.0)->add(-1))^10->sink - float           │ PASS │ 10  │    48  │   8k /  70k = 11.2% │   4k /  10M =  0.0% │   38.1  │  12 ms │  13 ms │ 215 us │  13 ms │  13 ms │     127 ms │  78.8M │
│ merged src->(mult(2.0)->div(2.0)->add(-1))^10->sink - int             │ PASS │ 10  │    54  │   8k /  89k =  9.4% │   5k /  10M =  0.0% │   48.1  │  15 ms │  15 ms │  22 us │  15 ms │  15 ms │     150 ms │  66.5M │
│ runtime   src->sink overhead                                          │ PASS │ 10  │     5  │  32k /   1M =  2.3% │   2k /  20M =  0.0% │   13.0  │   1 ms │   1 ms │  28 us │   1 ms │   1 ms │      13 ms │ 781.8M │
│ runtime   src->copy->sink                                             │ PASS │ 10  │     6  │ 132k /   4M =  3.6% │   2k /  20M =  0.0% │   13.4  │   1 ms │   1 ms │  35 us │   1 ms │   1 ms │      14 ms │ 736.7M │
│ runtime   src->copy^10->sink                                          │ PASS │ 10  │     8  │   4M /  24M = 17.9% │   8k /  24M =  0.0% │   17.0  │   2 ms │   2 ms │ 234 us │   2 ms │   3 ms │      22 ms │ 447.8M │
│ runtime   src(N=1024)->b1(N≤128)->b2(N=1024)->b3(N=32...128)->sink    │ PASS │ 10  │    10  │ 467k /   6M =  8.0% │ 154k /  32M =  0.5% │   20.2  │   2 ms │   2 ms │  32 us │   2 ms │   2 ms │      23 ms │ 441.9M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - float             │ PASS │ 10  │     7  │ 261k /   9M =  2.8% │   4k /  28M =  0.0% │   17.6  │   2 ms │   2 ms │  57 us │   2 ms │   2 ms │      17 ms │ 575.0M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - int               │ PASS │ 10  │    12  │ 366k /  10M =  3.8% │   5k /  35M =  0.0% │   24.6  │   3 ms │   3 ms │  55 us │   3 ms │   3 ms │      30 ms │ 328.4M │
│ runtime   src->(mult(2.0)->div(2.0)->add(-1))^10->sink - float        │ PASS │ 10  │    21  │   2M /  75M =  2.3% │  16k /  96M =  0.0% │   58.5  │   6 ms │   6 ms │ 312 us │   6 ms │   7 ms │      59 ms │ 169.4M │
│ runtime   src->(mult(2.0)->div(2.0)->add(-1))^10->sink - int          │ PASS │ 10  │    71  │   3M /  75M =  4.1% │  20k / 171M =  0.0% │  128.6  │  19 ms │  19 ms │ 316 us │  19 ms │  20 ms │     189 ms │  52.9M │
│ runtime   src->mult(2.0)->mult(0.5)->add(-1)->sink (SIMD)             │ PASS │ 10  │     5  │ 328k /  10M =  3.4% │   4k /  28M =  0.0% │   17.6  │   2 ms │   2 ms │  44 us │   2 ms │   2 ms │      16 ms │ 619.4M │
│ runtime   src->(mult(2.0)->mult(0.5)->add(-1))^10->sink (SIMD)        │ PASS │ 10  │    16  │   3M /  79M =  3.2% │  19k /  96M =  0.0% │   58.5  │   4 ms │   4 ms │ 350 us │   4 ms │   5 ms │      44 ms │ 226.1M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - float single      │ PASS │ 10  │     6  │ 260k /   9M =  2.8% │   4k /  28M =  0.0% │   17.6  │   2 ms │   2 ms │  52 us │   2 ms │   2 ms │      17 ms │ 573.2M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - int single        │ PASS │ 10  │    14  │ 233k /   9M =  2.5% │   4k /  35M =  0.0% │   24.6  │   3 ms │   3 ms │  51 us │   3 ms │   3 ms │      30 ms │ 328.7M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - float bulk        │ PASS │ 10  │     6  │ 225k /   9M =  2.4% │   4k /  28M =  0.0% │   17.6  │   2 ms │   2 ms │  56 us │   2 ms │   2 ms │      17 ms │ 583.8M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - int bulk          │ PASS │ 10  │    12  │ 240k /   9M =  2.7% │   4k /  35M =  0.0% │   24.6  │   3 ms │   3 ms │  58 us │   3 ms │   3 ms │      31 ms │ 326.1M │
└───────────────────────────────────────────────────────────────────────┴──────┴─────┴────────┴─────────────────────┴─────────────────────┴─────────┴────────┴────────┴────────┴────────┴────────┴────────────┴────────┘
```
### PR#50 - bm_case1

```
┌──────────────────────────────benchmark:───────────────────────────────┬──────┬─#N──┬─CTX-SW─┬──CPU cache misses───┬──CPU branch misses──┬─<CPU-I>─┬──min───┬──mean──┬─stddev─┬─median─┬──max───┬─total time─┬─ops/s──┐
│ merged src->sink work                                                 │ PASS │ 10  │     1  │   3k /   7k = 42.4% │ 792  / 322k =  0.2% │  348.0m │ 323 us │ 538 us │ 237 us │ 457 us │ 895 us │       5 ms │   1.9G │
│ merged src->copy->sink                                                │ PASS │ 10  │     2  │   1k /   5k = 29.6% │ 416  / 321k =  0.1% │  784.9m │ 342 us │ 358 us │  11 us │ 365 us │ 374 us │       4 ms │   2.8G │
│ merged src->copy->sink work                                           │ PASS │ 10  │     1  │   1k /   4k = 33.5% │ 264  / 319k =  0.1% │  878.2m │ 337 us │ 338 us │   3 us │ 338 us │ 346 us │       3 ms │   3.0G │
│ merged src->copy^10->sink                                             │ PASS │ 10  │     1  │ 738  /   3k = 23.3% │ 265  / 319k =  0.1% │  784.3m │ 318 us │ 325 us │   7 us │ 326 us │ 336 us │       3 ms │   3.1G │
│ merged src(N=1024)->b1(N≤128)->b2(N=1024)->b3(N=32...128)->sink       │ PASS │ 10  │     1  │   2k /   6k = 42.2% │ 411  / 321k =  0.1% │  285.2m │ 289 us │ 311 us │  15 us │ 318 us │ 325 us │       3 ms │   3.2G │
│ merged src->mult(2.0)->divide(2.0)->add(-1)->sink - float             │ PASS │ 10  │     3  │   6k /  12k = 45.4% │   1k / 953k =  0.1% │    1.9  │ 532 us │ 553 us │  39 us │ 540 us │ 668 us │       6 ms │   1.8G │
│ merged src->mult(2.0)->divide(2.0)->add(-1)->sink - int               │ PASS │ 10  │     4  │   5k /  13k = 35.9% │ 897  /   6M =  0.0% │    8.8  │   1 ms │   1 ms │  10 us │   1 ms │   1 ms │      11 ms │ 916.2M │
│ merged src->(mult(2.0)->div(2.0)->add(-1))^10->sink - float           │ PASS │ 10  │    10  │   2k /  16k = 11.3% │ 980  /   7M =  0.0% │   16.3  │   2 ms │   2 ms │  15 us │   2 ms │   3 ms │      25 ms │ 401.6M │
│ merged src->(mult(2.0)->div(2.0)->add(-1))^10->sink - int             │ PASS │ 10  │    27  │   4k /  43k =  9.1% │   2k /  57M =  0.0% │   84.4  │   7 ms │   7 ms │  34 us │   7 ms │   7 ms │      71 ms │ 140.7M │
│ runtime   src->sink overhead                                          │ PASS │ 10  │     2  │  28k /   2M =  1.5% │   2k /  11M =  0.0% │    7.9  │ 825 us │ 839 us │  22 us │ 830 us │ 901 us │       8 ms │   1.2G │
│ runtime   src->copy->sink                                             │ PASS │ 10  │     4  │ 128k /   3M =  4.1% │   2k /  11M =  0.0% │    8.3  │ 895 us │ 920 us │  31 us │ 909 us │   1 ms │       9 ms │   1.1G │
│ runtime   src->copy^10->sink                                          │ PASS │ 10  │     7  │   4M /  24M = 18.5% │   8k /  14M =  0.1% │   11.8  │   2 ms │   2 ms │ 129 us │   2 ms │   2 ms │      18 ms │ 544.9M │
│ runtime   src(N=1024)->b1(N≤128)->b2(N=1024)->b3(N=32...128)->sink    │ PASS │ 10  │     7  │ 462k /   6M =  7.9% │  76k /  23M =  0.3% │   15.1  │   2 ms │   2 ms │  39 us │   2 ms │   2 ms │      18 ms │ 541.4M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - float             │ PASS │ 10  │     6  │ 223k /  10M =  2.2% │   4k /  13M =  0.0% │   11.2  │   1 ms │   1 ms │  52 us │   1 ms │   1 ms │      13 ms │ 786.7M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - int               │ PASS │ 10  │     7  │ 188k /   9M =  2.0% │   4k /  13M =  0.0% │   17.0  │   2 ms │   2 ms │  58 us │   2 ms │   2 ms │      18 ms │ 543.7M │
│ runtime   src->(mult(2.0)->div(2.0)->add(-1))^10->sink - float        │ PASS │ 10  │    20  │   2M /  78M =  2.4% │  16k /  30M =  0.1% │   40.9  │   5 ms │   5 ms │ 291 us │   5 ms │   6 ms │      50 ms │ 199.1M │
│ runtime   src->(mult(2.0)->div(2.0)->add(-1))^10->sink - int          │ PASS │ 10  │    42  │   1M /  75M =  1.9% │  18k /  31M =  0.1% │   98.4  │  11 ms │  11 ms │ 282 us │  11 ms │  12 ms │     108 ms │  92.7M │
│ runtime   src->mult(2.0)->mult(0.5)->add(-1)->sink (SIMD)             │ PASS │ 10  │     5  │ 247k /  10M =  2.4% │   4k /  18M =  0.0% │   12.4  │   1 ms │   1 ms │  32 us │   1 ms │   1 ms │      12 ms │ 826.3M │
│ runtime   src->(mult(2.0)->mult(0.5)->add(-1))^10->sink (SIMD)        │ PASS │ 10  │    15  │   2M /  79M =  3.0% │  19k /  87M =  0.0% │   53.4  │   4 ms │   4 ms │ 395 us │   4 ms │   5 ms │      41 ms │ 242.5M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - float single      │ PASS │ 10  │     4  │ 234k /  10M =  2.4% │   4k /  13M =  0.0% │   11.2  │   1 ms │   1 ms │  45 us │   1 ms │   1 ms │      13 ms │ 781.9M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - int single        │ PASS │ 10  │     7  │ 185k /   9M =  2.0% │   4k /  13M =  0.0% │   17.0  │   2 ms │   2 ms │  52 us │   2 ms │   2 ms │      19 ms │ 539.7M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - float bulk        │ PASS │ 10  │     5  │ 233k /  10M =  2.3% │   4k /  18M =  0.0% │   12.5  │   1 ms │   1 ms │  50 us │   1 ms │   1 ms │      13 ms │ 775.4M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - int bulk          │ PASS │ 10  │    11  │ 225k /  10M =  2.3% │   4k /  26M =  0.0% │   19.5  │   3 ms │   3 ms │  61 us │   3 ms │   3 ms │      26 ms │ 380.6M │
└───────────────────────────────────────────────────────────────────────┴──────┴─────┴────────┴─────────────────────┴─────────────────────┴─────────┴────────┴────────┴────────┴────────┴────────┴────────────┴────────┘
```
### PR_50 - bm_case1_nosimd

```
┌──────────────────────────────benchmark:───────────────────────────────┬──────┬─#N──┬─CTX-SW─┬──CPU cache misses───┬──CPU branch misses──┬─<CPU-I>─┬──min───┬──mean──┬─stddev─┬─median─┬──max───┬─total time─┬─ops/s──┐
│ merged src->sink work                                                 │ PASS │ 10  │     1  │   2k /   6k = 33.6% │ 585  /  10M =  0.0% │    8.0  │ 497 us │ 538 us │  90 us │ 513 us │ 806 us │       5 ms │   1.9G │
│ merged src->copy->sink                                                │ PASS │ 10  │     1  │   1k /   4k = 29.6% │ 369  /  10M =  0.0% │    7.0  │ 497 us │ 513 us │   9 us │ 515 us │ 528 us │       5 ms │   2.0G │
│ merged src->copy->sink work                                           │ PASS │ 10  │     1  │   1k /   4k = 24.4% │ 456  /  10M =  0.0% │    8.0  │ 513 us │ 516 us │   4 us │ 515 us │ 525 us │       5 ms │   1.9G │
│ merged src->copy^10->sink                                             │ PASS │ 10  │     1  │ 792  /   4k = 18.6% │ 336  /  10M =  0.0% │    7.0  │ 497 us │ 504 us │   7 us │ 505 us │ 516 us │       5 ms │   2.0G │
│ merged src(N=1024)->b1(N≤128)->b2(N=1024)->b3(N=32...128)->sink       │ PASS │ 10  │     2  │   2k /   7k = 22.4% │ 551  /  10M =  0.0% │    7.0  │ 497 us │ 512 us │  16 us │ 513 us │ 557 us │       5 ms │   2.0G │
│ merged src->mult(2.0)->divide(2.0)->add(-1)->sink - float             │ PASS │ 10  │     5  │   3k /  12k = 24.0% │ 797  /  10M =  0.0% │   12.0  │   1 ms │   1 ms │  14 us │   1 ms │   1 ms │      11 ms │ 920.9M │
│ merged src->mult(2.0)->divide(2.0)->add(-1)->sink - int               │ PASS │ 10  │     7  │   2k /  13k = 17.3% │ 905  /  10M =  0.0% │   13.0  │   1 ms │   2 ms │  23 us │   2 ms │   2 ms │      15 ms │ 657.7M │
│ merged src->(mult(2.0)->div(2.0)->add(-1))^10->sink - float           │ PASS │ 10  │    45  │   9k /  75k = 11.5% │   4k /  10M =  0.0% │   38.1  │  12 ms │  12 ms │ 325 us │  12 ms │  13 ms │     125 ms │  80.2M │
│ merged src->(mult(2.0)->div(2.0)->add(-1))^10->sink - int             │ PASS │ 10  │    56  │  12k /  96k = 12.7% │   5k /  10M =  0.1% │   48.1  │  15 ms │  15 ms │  42 us │  15 ms │  15 ms │     151 ms │  66.4M │
│ runtime   src->sink overhead                                          │ PASS │ 10  │     5  │  29k /   1M =  2.2% │   2k /  20M =  0.0% │   13.0  │   1 ms │   1 ms │  23 us │   1 ms │   1 ms │      13 ms │ 784.6M │
│ runtime   src->copy->sink                                             │ PASS │ 10  │     6  │ 135k /   4M =  3.7% │   3k /  20M =  0.0% │   13.4  │   1 ms │   1 ms │  34 us │   1 ms │   1 ms │      14 ms │ 733.9M │
│ runtime   src->copy^10->sink                                          │ PASS │ 10  │     9  │   4M /  24M = 17.5% │   8k /  24M =  0.0% │   17.0  │   2 ms │   2 ms │ 174 us │   2 ms │   3 ms │      22 ms │ 445.5M │
│ runtime   src(N=1024)->b1(N≤128)->b2(N=1024)->b3(N=32...128)->sink    │ PASS │ 10  │     9  │ 432k /   6M =  7.6% │ 154k /  32M =  0.5% │   20.1  │   2 ms │   2 ms │ 151 us │   2 ms │   3 ms │      24 ms │ 417.7M │
│ runtime   src->mult(2.0)->div(2.0)->add(-1)->sink - float             │ PASS │ 10  │     7  │ 264k /   9M =  2.8% │   4k /  28M =  0.0% │   17.6  │   2 ms │   2 ms │  40 us │   2 ms │   2 ms │      18 ms │ 560.9M │
…